WIP2

+++ sane/core/app.hpp
@@ -34,8 +34,8 @@ namespace Sane
         void PopOverlay(Layer* layer);
 
     private:
-        const size_t WIDTH = 1280;
-        const size_t HEIGHT = 720;
+        const int32_t WIDTH = 1280;
+        const int32_t HEIGHT = 720;
 
         Events::Queue& evt_queue_;
         Layers::Stack layers_;
